requirements.txt
Flask>=2.0.0
python-dotenv>=0.19.0
supabase>=2.0.0
Flask-Cors>=3.0.0